backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pyjwt
passlib[bcrypt]
psycopg2-binary
websockets
